requirements.txt
streamlit
moviepy
pydub
pyttsx3
python-dotenv 
pyrubberband
